fix(show-me): use IPC for clipboard example
Electron v20.0.0 enabled renderer sandboxing by default
and as a result clipboard can't be used in preload.js anymore
