+++ static/show-me/clipboard/index.html
@@ -1,6 +1,8 @@
 <!DOCTYPE html>
 <html lang="en">
   <head>
+    <meta charset="UTF-8">
+    <meta http-equiv="Content-Security-Policy" content="default-src 'self'; script-src 'self'">
     <title>Clipboard</title>
   </head>
   <body>
@@ -9,5 +11,6 @@
     <button id="paste">Paste into TextArea</button>
     <br />
     <textarea></textarea>
+    <script src="./renderer.js"></script>
   </body>
 </html>
